RemoteStartStopController: moved the task related methods to the new  TaskRestController class
