+++ src/main/java/de/rwth/idsg/steve/web/api/TaskRestController.java
@@ -1,0 +1,106 @@
+/*
+ * SteVe - SteckdosenVerwaltung - https://github.com/steve-community/steve
+ * Copyright (C) 2013-2023 SteVe Community Team
+ * All Rights Reserved.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+package de.rwth.idsg.steve.web.api;
+
+import de.rwth.idsg.steve.repository.TaskStore;
+
+import org.springframework.beans.factory.annotation.Autowired;
+
+import de.rwth.idsg.steve.web.api.ApiControllerAdvice.ApiErrorResponse;
+import de.rwth.idsg.steve.web.api.dto.ApiTaskInfo;
+import de.rwth.idsg.steve.web.api.dto.ApiTaskList;
+import de.rwth.idsg.steve.web.api.exception.BadRequestException;
+import io.swagger.annotations.ApiResponse;
+import io.swagger.annotations.ApiResponses;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.MediaType;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.RestController;
+
+import javax.validation.Valid;
+
+/**
+ * @author fnkbsi
+ * @since 18.10.2023
+ */
+@Slf4j
+@RestController
+@RequestMapping(value = "/api/v1/remote", produces = MediaType.APPLICATION_JSON_VALUE)
+@RequiredArgsConstructor
+public class TaskRestController {
+
+    @Autowired private TaskStore taskStore;
+
+    // -------------------------------------------------------------------------
+    // Helpers
+    // -------------------------------------------------------------------------
+
+    private ApiTaskList getTaskList() {
+        ApiTaskList lsTasks = new ApiTaskList();
+        lsTasks.setTasks(taskStore.getOverview());
+        return lsTasks;
+    }
+
+    // -------------------------------------------------------------------------
+    // Http methods (GET)
+    // -------------------------------------------------------------------------
+
+   @ApiResponses(value = {
+        @ApiResponse(code = 200, message = "OK"),
+        @ApiResponse(code = 400, message = "Bad Request", response = ApiErrorResponse.class),
+        @ApiResponse(code = 401, message = "Unauthorized", response = ApiErrorResponse.class),
+        @ApiResponse(code = 500, message = "Internal Server Error", response = ApiErrorResponse.class)}
+    )
+    @GetMapping(value = "taskoverview")
+    @ResponseBody
+    public ApiTaskList getOverview() {
+        return getTaskList();
+    }
+
+
+    @ApiResponses(value = {
+        @ApiResponse(code = 200, message = "OK"),
+        @ApiResponse(code = 400, message = "Bad Request", response = ApiErrorResponse.class),
+        @ApiResponse(code = 401, message = "Unauthorized", response = ApiErrorResponse.class),
+        @ApiResponse(code = 500, message = "Internal Server Error", response = ApiErrorResponse.class)}
+    )
+    @PostMapping(value = "clearfinishedtasks")
+    @ResponseBody
+    public ApiTaskList clearFinished() {
+        taskStore.clearFinished();
+        return getTaskList();
+    }
+
+    @ApiResponses(value = {
+        @ApiResponse(code = 200, message = "OK"),
+        @ApiResponse(code = 400, message = "Bad Request", response = ApiErrorResponse.class),
+        @ApiResponse(code = 401, message = "Unauthorized", response = ApiErrorResponse.class),
+        @ApiResponse(code = 500, message = "Internal Server Error", response = ApiErrorResponse.class)}
+    )
+    @GetMapping(value = "task")
+    @ResponseBody
+    public ApiTaskInfo getTaskDetails(@Valid Integer taskId) {
+        ApiTaskInfo taskInfo = new ApiTaskInfo(taskId, taskStore.get(taskId));
+        return taskInfo;
+    }
+}
